Move jwt_conversion code to separate file
